refactor(schema-registry): share version parsing
